requirements.txt
aiosmtpd>=1.4.4.post2,<2
kubernetes>=28.1.0,<29
google-api-python-client>=2.103.0,<4
google-auth-httplib2>=0.1.1,<2
google-auth-oauthlib>=1.1.0,<3
mailjet_rest >=1.3.4,<2
